Issue #4587: Moved all resources for UncommentedMain check
