+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/misc/uncommentedmain/InputUncommentedMain3.java
@@ -2,13 +2,12 @@
 // Test case file for checkstyle.
 // Created: 2003
 ////////////////////////////////////////////////////////////////////////////////
-package com.puppycrawl.tools.checkstyle.checks;
+package com.puppycrawl.tools.checkstyle.checks.misc.uncommentedmain;
 
-class InputUncommentedMainTest4
-{
-    // one more uncommented main
-    public static void main(int[] args)
+class oneMoreClass {
+    
+    public static void anyWrongMethodName(String[] args)
     {
-        System.identityHashCode("test1.main()");
+        System.identityHashCode("InputUncommentedMain.main()");
     }
 }
