+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/misc/uncommentedmain/InputUncommentedMain4.java
@@ -2,12 +2,13 @@
 // Test case file for checkstyle.
 // Created: 2003
 ////////////////////////////////////////////////////////////////////////////////
-package com.puppycrawl.tools.checkstyle.checks;
+package com.puppycrawl.tools.checkstyle.checks.misc.uncommentedmain;
 
-class oneMoreClass {
-    
-    public static void anyWrongMethodName(String[] args)
+class InputUncommentedMainTest4
+{
+    // one more uncommented main
+    public static void main(int[] args)
     {
-        System.identityHashCode("InputUncommentedMain.main()");
+        System.identityHashCode("test1.main()");
     }
 }
